+++ Pixel Art Core/Canvas/PXCanvas_Layers.h
@@ -13,7 +13,7 @@
 - (PXLayer *) activeLayer;
 - (void)activateLayer:(PXLayer *) aLayer;
 - (NSArray *) layers;
-- (int)indexOfLayer:(PXLayer *) aLayer;
+- (NSUInteger)indexOfLayer:(PXLayer *) aLayer;
 
 - (void)setLayers:(NSArray *) newLayers;
 - (void)setLayers:(NSArray*)layers fromLayers:(NSArray *)oldLayers;
@@ -22,19 +22,19 @@
 
 - (void)addLayer:(PXLayer *) aLayer suppressingNotification:(BOOL)suppress;
 - (void)addLayer:(PXLayer *)aLayer;
-- (void)insertLayer:(PXLayer *) aLayer atIndex:(int)index suppressingNotification:(BOOL)suppress;
-- (void)insertLayer:(PXLayer *) aLayer atIndex:(int)index;
+- (void)insertLayer:(PXLayer *) aLayer atIndex:(NSUInteger)index suppressingNotification:(BOOL)suppress;
+- (void)insertLayer:(PXLayer *) aLayer atIndex:(NSUInteger)index;
 - (void)removeLayer: (PXLayer*) aLayer;
 - (void)removeLayer: (PXLayer*) aLayer suppressingNotification:(BOOL)suppress;
-- (void)removeLayerAtIndex:(int)index;
-- (void)removeLayerAtIndex:(int)index suppressingNotification:(BOOL)suppress;
+- (void)removeLayerAtIndex:(NSUInteger)index;
+- (void)removeLayerAtIndex:(NSUInteger)index suppressingNotification:(BOOL)suppress;
 
-- (void)moveLayer:(PXLayer *)aLayer toIndex:(int)anIndex;
+- (void)moveLayer:(PXLayer *)aLayer toIndex:(NSUInteger)anIndex;
 - (void)layersChanged;
 
 - (void)rotateLayer:(PXLayer *)layer byDegrees:(int)degrees;
 
-- (void)duplicateLayerAtIndex:(unsigned)index;
+- (void)duplicateLayerAtIndex:(NSUInteger)index;
 - (void)flipLayerHorizontally:aLayer;
 - (void)flipLayerVertically:aLayer;
