+++ Pixen Application/Tools/Tools/PXMoveTool.m
@@ -180,7 +180,7 @@ fromCanvasController:(PXCanvasController *) controller
 			[copyLayer finalizeMotion];
 		}
 		selectedRect = lastSelectedRect;
-		int index = [[canvas layers] indexOfObject:moveLayer];
+		NSUInteger index = [[canvas layers] indexOfObject:moveLayer];
 		[canvas mergeDownLayer:moveLayer];
 		[moveLayer release];
 		if(isCopying && copyLayer != nil)
